Allow any goal unit; derive progress from the task's own stated quantity
Removes the hardcoded list of "uncountable" units (hours, sessions,
classes, ...) per product direction: the user confirms task completion
themselves, so any goal with a target/unit should be eligible for
credit. Replaces it with text-driven extraction: the backend parses an
explicit "<quantity> <unit>" phrase out of the task's own title/
description and converts it into the goal's unit (e.g. "15 minutes" ->
0.25 hours, "one session" -> 1 session), rather than trusting the AI's
claimed amount directly. The AI's claim is still cross-checked against
the derived amount and rejected on meaningful disagreement.

This makes progress fractional, so target/progress widen from int to
float/num across both backend (GoalContext, TaskSuggestion, schemas)
and frontend (MonthlyGoal, TaskSuggestion, goal repository). Requires
widening the live monthly_goals.target/progress Supabase columns from
integer to numeric (safe, lossless for existing whole-number values).

+++ backend/app/domain/goal_context.py
@@ -12,6 +12,6 @@ class GoalContext:
 
     id: str
     title: str
-    target: int
+    target: float
     unit: str
-    progress: int
+    progress: float
